removed type cast to uintptr_t

+++ src/mpg/map_var.c
@@ -86,7 +86,7 @@ void  MPG_Cart_varfree (int dim, int *start, int *end, int *pad,
 
   for (i=0; i<ptr_list_len & (char *)var != ptr_list[i]; i++);
   if (i == ptr_list_len) {
-    fprintf (stderr, "Cannot find record of allocation at address %#lx\n",(uintptr_t)var);
+    fprintf (stderr, "Cannot find record of allocation at address %#lx\n",(unsigned long int)var);
     MPI_Abort (MPI_COMM_WORLD, 0);
   }
